+++ lib/redgraph/graph/edge_methods.rb
@@ -6,13 +6,14 @@ module Redgraph
       # Adds an edge. If successul it returns the created object, otherwise false
       #
       def add_edge(edge)
-        result = _query("MATCH (src), (dest)
-                        WHERE ID(src) = #{edge.src.id} AND ID(dest) = #{edge.dest.id}
-                        CREATE (src)-[e:`#{edge.type}` #{quote_hash(edge.properties)}]->(dest) RETURN ID(e)")
-        return false if result.stats[:relationships_created] != 1
-        id = result.resultset.first["ID(e)"]
-        edge.id = id
-        edge
+        merge_or_add_edge(edge, :create)
+      end
+
+      # Merges (creates an edge unless one with the same type and properties exists) an edge.
+      # If successul it returns the object, otherwise false
+      #
+      def merge_edge(edge)
+        merge_or_add_edge(edge, :merge)
       end
 
       # Finds edges. Options:
@@ -74,6 +75,17 @@ module Redgraph
           edge.type = get_relationship_type(type_id)
           edge.properties = attrs
         end
+      end
+
+      def merge_or_add_edge(edge, verb = :create)
+        verb = verb == :create ? "CREATE" : "MERGE"
+        result = _query("MATCH (src), (dest)
+                        WHERE ID(src) = #{edge.src.id} AND ID(dest) = #{edge.dest.id}
+                        #{verb} (src)-[e:`#{edge.type}` #{quote_hash(edge.properties)}]->(dest) RETURN ID(e)")
+        return false if result.stats[:relationships_created] != 1
+        id = result.resultset.first["ID(e)"]
+        edge.id = id
+        edge
       end
     end
   end