+++ lib/redgraph/node_model.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 require_relative 'node_model/class_methods'
-require_relative 'node_model/registry'
 
 module Redgraph
   # This mixin allows you to use an interface similar to ActiveRecord
@@ -18,18 +17,17 @@ module Redgraph
   # john = Actor.find(123)
   # total = Actor.count
   #
-  # When you include this module (or subclass a class with it) then the label-class registry is
-  # updated.
+  # When you create a record it will automatically set the _type property with the class name.
+  # This allows reifying the node into the corresponding NodeModel class.
   #
   module NodeModel
     extend ActiveSupport::Concern
 
     included do |base|
-      Registry.register_node_model(base)
-
       @attribute_names = [:id]
 
       attr_accessor :id
+      attr_accessor :_type
 
       class << self
         attr_reader :attribute_names
@@ -46,13 +44,12 @@ module Redgraph
           super
           subclass.instance_variable_set(:@attribute_names, @attribute_names.dup)
           subclass.instance_variable_set(:@graph, @graph.dup)
-          Registry.register_node_model(subclass)
         end
       end
     end
 
     def initialize(**args)
-      absent_attributes = args.keys.map(&:to_sym) - self.class.attribute_names
+      absent_attributes = args.keys.map(&:to_sym) - self.class.attribute_names - [:_type]
 
       if absent_attributes.any?
         raise ArgumentError, "Unknown attribute #{absent_attributes}"
@@ -108,7 +105,8 @@ module Redgraph
     end
 
     def to_node
-      Redgraph::Node.new(id: id, label: label, properties: attributes.except(:id))
+      props = attributes.except(:id).merge(_type: self.class.name)
+      Redgraph::Node.new(id: id, label: label, properties: props)
     end
 
     # Converts a Node object into NodeModel